refactor: transaction form
