+++ src/constants/forms.ts
@@ -1,0 +1,4 @@
+export enum FormType {
+  ADD = "add",
+  EDIT = "edit",
+}
